Move 'office' and 'program' fixtures to conftest.py
